refactor(embedding_compute): extract providers into adapter pattern
- Split embedding_compute.py (1444→403 lines) into registry + dispatch
- New providers/ package with one module per backend:
  - sentence_transformers.py (487 lines) — ST models + hardware optimizations
  - openai.py (136 lines) — OpenAI-compatible API
  - mlx.py (82 lines) — Apple Silicon MLX
  - ollama.py (335 lines) — Ollama local server
  - gemini.py (89 lines) — Google Gemini API
- Add _init_providers() lazy-import + _providers registry dict
- Add register_provider(name, fn) for third-party backends
- Full backward compat: old function names (compute_embeddings_sentence_transformers
  etc.) kept as module-level aliases replaced at provider init
- Shared utilities (_model_cache, get_model_token_limit, truncate_to_token_limit,
  _query_ollama_context_limit, _query_lmstudio_context_limit) stay in main module
- Adding a new provider now requires only a new file + 1 registry line
